Use in-flight elements queue instead of thread pool saturation to propagate backpressure
Prior to this commit, we were using the size of the thread pool as a means to propagate backpressure
upstream when the thread pool was fully occupied. This was a leftover from the original design, where
the ExactlyOnceCollector provided to UDF invokes directly emit collected output.

Since in the latest approach, we added a separate OutputEmitter to consume the queue and do the actual
emitting, the backpressure can only be correctly propagated from there. This change therefore includes
capping the size of the queue, and logic to block threads adding elements until there is space.