[FLINK-6190] [core] Reconfiguration for enum TypeSerializers
This commit adds reconfiguration implementation for the EnumSerializer
and Scala's EnumValueSerializer. Now, the enum serializers are capable
of being reconfigured so that the ordering of enum constants remains the
same as its preceding enum serializer. This allows the new enum
serializer's enum constants to have different ordinals, as well as new enum
constants. The only limitation is that previous enum constants cannot be
removed.