[FLINK-6190] [core] Reconfiguration for Kryo-related TypeSerializers
This commit adds implementation for Kryo-related serializers, most
notably the KryoSerializer, as well as AvroSerializer and
ValueSerializer, which all use Kryo and has Kryo registrations as part
of their serializer parameters.

Reconfiguration for Kryo-related serializers mainly deals with
reordering registrations so that new serializers use the same
registration ids for the same classes as the preceding serializer.